requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
#Pillow==7.1.1  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6 --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
redis==3.4.1 # https://github.com/andymccurdy/redis-py
celery==4.4.2  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==2.0.0  # https://github.com/celery/django-celery-beat
sorl-thumbnail==12.5.0 #处理略缩图的package

# Django
# ------------------------------------------------------------------------------
django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4  # https://github.com/django-compressor/django-compressor
django-redis==4.11.0  # https://github.com/niwinz/django-redis
django-taggit==1.1.0
django-mdeditor==0.1.14
django-markdownx==3.0.1
django-contrib-comments==1.9.2


